+++ packages/fabric/editor-common/src/utils/validator.ts
@@ -355,6 +355,28 @@ export const getValidNode = (
         }
         break;
       }
+      case 'inlineExtension':
+      case 'extension': {
+        if (attrs && attrs.extensionType && attrs.extensionKey) {
+          return {
+            type,
+            attrs,
+          };
+        }
+        break;
+      }
+      case 'bodiedExtension': {
+        if (attrs && attrs.extensionType && attrs.extensionKey && content) {
+          // we want to keep the oringinal Atlassian Document Content here
+          attrs.originalContent = JSON.parse(JSON.stringify(content));
+          return {
+            type,
+            attrs,
+            content,
+          };
+        }
+        break;
+      }
       case 'hardBreak': {
         return {
           type,