+++ packages/fabric/media-avatar-picker/src/image-cropper/index.tsx
@@ -1,8 +1,14 @@
 /* tslint:disable:variable-name */
 import * as React from 'react';
-import {Component} from 'react';
-import {CircularMask, Container, DragOverlay, RectMask, Image} from './styled';
-import {isImageRemote} from '@atlaskit/media-core';
+import { Component } from 'react';
+import {
+  CircularMask,
+  Container,
+  DragOverlay,
+  RectMask,
+  Image,
+} from './styled';
+import { isImageRemote } from '@atlaskit/media-core';
 
 export interface LoadParameters {
   export: () => string;
@@ -39,19 +45,19 @@ export class ImageCropper extends Component<ImageCropperProp, {}> {
   componentDidMount() {
     if (this.props.onLoad) {
       this.props.onLoad({
-        export: this.export
+        export: this.export,
       });
     }
   }
 
   onDragStarted = () => this.props.onDragStarted && this.props.onDragStarted();
 
-  onImageLoaded = (e) => {
+  onImageLoaded = e => {
     if (this.props.onImageSize) {
       this.props.onImageSize(e.target.naturalWidth, e.target.naturalHeight);
     }
     this.imageElement = e.target;
-  }
+  };
 
   render() {
     const {
@@ -59,12 +65,12 @@ export class ImageCropper extends Component<ImageCropperProp, {}> {
       containerSize,
       top,
       left,
-      imageSource
+      imageSource,
     } = this.props;
 
     const containerStyle = {
       width: `${containerSize}px`,
-      height: `${containerSize}px`
+      height: `${containerSize}px`,
     };
 
     const width = this.width ? `${this.width}px` : 'auto';
@@ -77,24 +83,31 @@ export class ImageCropper extends Component<ImageCropperProp, {}> {
     };
     const crossOrigin = isImageRemote(imageSource) ? 'anonymous' : undefined;
 
-    return <Container style={containerStyle}>
-      <Image crossOrigin={crossOrigin} src={imageSource} style={imageStyle} onLoad={this.onImageLoaded} />
-      {isCircularMask ? <CircularMask /> : <RectMask />}
-      <DragOverlay onMouseDown={this.onDragStarted} />
-    </Container>;
+    return (
+      <Container style={containerStyle}>
+        <Image
+          crossOrigin={crossOrigin}
+          src={imageSource}
+          style={imageStyle}
+          onLoad={this.onImageLoaded}
+        />
+        {isCircularMask ? <CircularMask /> : <RectMask />}
+        <DragOverlay onMouseDown={this.onDragStarted} />
+      </Container>
+    );
   }
 
   private get width() {
-    const {imageWidth, scale} = this.props;
+    const { imageWidth, scale } = this.props;
 
     return imageWidth ? imageWidth * (scale || defaultScale) : 0;
   }
 
-  export = () : string => {
+  export = (): string => {
     let imageData = '';
     const containerPadding = 20;
     const canvas = document.createElement('canvas');
-    const {top, left, scale, containerSize} = this.props;
+    const { top, left, scale, containerSize } = this.props;
     const size = containerSize || 0;
     const scaleWithDefault = scale || 1;
     const destinationSize = Math.max(size - containerPadding * 2, 0);
@@ -110,12 +123,20 @@ export class ImageCropper extends Component<ImageCropperProp, {}> {
       const sourceWidth = destinationSize / scaleWithDefault;
       const sourceHeight = destinationSize / scaleWithDefault;
 
-      context.drawImage(this.imageElement,
-        sourceLeft, sourceTop, sourceWidth, sourceHeight,
-        0, 0, destinationSize, destinationSize);
+      context.drawImage(
+        this.imageElement,
+        sourceLeft,
+        sourceTop,
+        sourceWidth,
+        sourceHeight,
+        0,
+        0,
+        destinationSize,
+        destinationSize,
+      );
       imageData = canvas.toDataURL();
     }
 
     return imageData;
-  }
+  };
 }