+++ packages/elements/breadcrumbs/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @atlaskit/breadcrumbs
 
+## 3.1.7
+
 ## 3.1.6
 
 ## 3.1.5