+++ packages/utils/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @atlaskit/docs
 
+## 2.1.1
+- [patch] Convert function parameters [f6c5a21](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f6c5a21)
+- [patch] Convert function parameters [f6c5a21](https://bitbucket.org/atlassian/atlaskit-mk-2/commits/f6c5a21)
+
 ## 2.1.0